Time: N/A (0%), Space: N/A (0%) - LeetHub

+++ 0050-powx-n/0050-powx-n.cpp
@@ -1,13 +1,13 @@
 class Solution {
 public:
     double myPow(double x, int n) {
-        long binform = n;
+        long long binform = n;
         if(n < 0)
         {
             x = 1/x;
             binform = -binform;
         }
-        double ans=1;
+        double ans = 1.0;
         while(binform > 0)
         {
             if(binform & 1)
